recvlen must be signed.
git-svn-id: file:///var/www/svn.irssi.org/SVN/irssi/trunk@4974 dbcabf3a-b0e7-0310-adc4-f8d773084564

+++ src/fe-common/core/fe-exec.c
@@ -350,7 +350,7 @@ static void process_exec(PROCESS_REC *rec, const char *cmd)
 static void sig_exec_input_reader(PROCESS_REC *rec)
 {
         char tmpbuf[512], *str;
-        gsize recvlen;
+        int recvlen;
 	int ret;
 
 	g_return_if_fail(rec != NULL);
